Added ICC profile, undecorated first frame
Added ICC color profile for C740-14IML from:
https://www.notebookcheck.net/Lenovo-Yoga-C740-14IML-Convertible-Review-Top-tier-Looks-Mid-tier-Price.454741.0.html

Tried various methods of creating an inisial frame
with the "undecorated" frame parameter, to remove
fringes/borders on the initial EXWM workspace.

The fringes can be seen with (x-frame-geometry) in
the (external-border-size). The only thing that worked was modifying the
frame paramter in the doom config.el after the frame has been created by
emacsclient.